Correct TxFilterConfig on prover-client (#723)
* feat: update CL STF

* feat: create BatchTransition struct

* fix: update worker

* refactor: better handling of tx filter update

* wip Update BtcBlockspace operator to construct right TxFilterConfig

* fixup

* fix spellings

* ignore unit test

* refactor: use explicit struct for each transition

* refactor: add comments

* refactor: extract into a separate function

* fix(tests): use non-empty L1Segment in L2Block

* fix: functional-tests

+++ crates/test-utils/src/l2.rs
@@ -139,8 +139,7 @@ pub fn get_genesis_chainstate() -> Chainstate {
     let btc_chain = BtcChainSegment::load();
     // Build the genesis block and genesis consensus states.
     let gblock = make_genesis_block(&params);
-    let pregenesis_mfs =
-        vec![btc_chain.get_block_manifest(params.rollup().genesis_l1_height as u32)];
+    let pregenesis_mfs = vec![btc_chain.get_block_manifest(params.rollup().genesis_l1_height)];
     make_genesis_chainstate(&gblock, pregenesis_mfs, &params)
 }
